Fix tests
No more server start, not needed for graphql
