+++ back/integration-tests/user.test.ts
@@ -1,19 +1,13 @@
-import { execute, closeServer, resetDatabase } from "./helper";
+import { execute, resetDatabase } from "./helper";
 import { prisma } from "../src/generated/prisma-client";
 import * as mailsHelper from "../src/common/mails.helper";
-import * as sub from "../src/subscriptions/index";
 
 // No mails
 const sendMailSpy = jest.spyOn(mailsHelper, "sendMail");
 sendMailSpy.mockImplementation(() => Promise.resolve());
 
-// No subscriptions
-const subSpy = jest.spyOn(sub, "initSubscriptions");
-subSpy.mockImplementation(() => null);
-
 describe("User endpoint", () => {
   afterAll(async () => {
-    await closeServer();
     await resetDatabase();
   });
